проведен рефакторинг кода
