+++ src/mock/task.js
@@ -1,9 +1,5 @@
-const getRandomInteger = (a = 0, b = 1) => {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-
-  return Math.floor(lower + Math.random() * (upper - lower + 1));
-};
+import {getRandomInteger} from '../utils';
+import {COLORS} from '../const';
 
 const generateTaskDescription = () => {
   const description = [`Изучить теорию`, `Сделать домашку`, `Пройти интенсив на соточку`];
@@ -41,10 +37,9 @@ const generateRepeatingTask = () => {
 };
 
 const generateTaskColor = () => {
-  const colors = [`black`, `yellow`, `blue`, `green`, `pink`];
-  const colorIndex = getRandomInteger(0, colors.length - 1);
+  const colorIndex = getRandomInteger(0, COLORS.length - 1);
 
-  return colors[colorIndex];
+  return COLORS[colorIndex];
 };
 
 export const generateTask = () => {
