+++ src/utils.js
@@ -1,0 +1,22 @@
+export const getRandomInteger = (a = 0, b = 1) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+
+  return Math.floor(lower + Math.random() * (upper - lower + 1));
+};
+
+export const isExpiredTask = dueDate => {
+  if (dueDate === null) {
+    return false;
+  }
+
+  const currentDate = new Date();
+  currentDate.setHours(23, 59, 59, 999);
+  return currentDate.getTime() > dueDate.getTime();
+};
+
+export const isRepeatingTask = repeating => {
+  return Object.values(repeating).some(Boolean);
+};
+
+export const humanizeTaskDueDate = dueDate => dueDate.toLocaleString('en-US', {day: 'numeric', month: 'long'});
